First commit to add this demo project.
This project is a "first draft" to work on some of the basic GUI
features and internal functionality of the software. Hopefully things
won't get too messy but this is a good place to write as a team and
later we can sit down to create our version of greenThink.

Erik